Fix paste auth

+++ install.bash
@@ -1,3 +1,9 @@
+if [ ! -e "~/.token/paste" ]; then
+    echo "Please get the paste token and paste to ~/.token/paste"
+    echo "https://nda.ya.ru/t/iYFw9Pty6o8qqc"
+    exit 1
+fi
+
 echo 'Install base apt packages...'
 sudo apt update
 sudo apt install -y zsh build-essentialsoftware-properties-common neovim
@@ -20,7 +26,19 @@ git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:
 echo 'Install cooler tools...'
 cargo install fd-find du-dust ripgrep git-delta bat hexyl exa zoxide
 pip install -U pip
-pip install -U httpx
+pip install -U httpie
+
+function getpaste() {
+    if [[ "$1" == https://nda.ya.ru* ]] || [[ "$1" == nda.ya.ru* ]]; then
+        paste_url="$(https https://nda.ya.ru/t/vhLhmMOk6o8piM -h | rg Location | sed 's/Location: //g;s/\r//g')"
+    else
+        paste_url="$1"
+    fi
+    https --follow --body "$paste_url" "Authorization:OAuth $(cat ~/.token/paste)"
+}
 
 echo 'Install yandex tools...'
-curl -L https://nda.ya.ru/t/Yxq2Yzsq4FQnvL | bash
+getpaste https://nda.ya.ru/t/Yxq2Yzsq4FQnvL | bash
+
+echo 'Install yandex zshrc...'
+getpaste https://nda.ya.ru/t/vhLhmMOk6o8piM > ~/.zshrc-yandex
